Completed tutorial, added addtl logic to Delete method of controller

+++ MvcMovie/Controllers/MoviesController.cs
@@ -25,9 +25,12 @@ namespace MvcMovie.Controllers
         //
         // GET: /Movies/Details/5
 
-        public ViewResult Details(int id)
+        public ActionResult Details(int id)
         {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+                return HttpNotFound();
+
             return View(movie);
         }
 
@@ -86,9 +89,13 @@ namespace MvcMovie.Controllers
         //
         // GET: /Movies/Delete/5
  
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int id = 0)
         {
             Movie movie = db.Movies.Find(id);
+
+            if (movie == null)
+                return HttpNotFound();
+
             return View(movie);
         }
 
@@ -96,9 +103,12 @@ namespace MvcMovie.Controllers
         // POST: /Movies/Delete/5
 
         [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id = 0)
         {            
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+                return HttpNotFound();
+
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
